Move new implementations to the proper files
