--- go/vt/vtgate/vcursor_background.go
@@ -1,31 +1,0 @@
-/*
-Copyright 2022 The Vitess Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package vtgate
-
-import (
-	"google.golang.org/protobuf/proto"
-
-	vtgatepb "vitess.io/vitess/go/vt/proto/vtgate"
-	"vitess.io/vitess/go/vt/vtgate/engine"
-)
-
-func (vc *vcursorImpl) BackgroundVCursor() engine.VCursor {
-	cloneCtx := vc.clone()
-	sess := proto.Clone(cloneCtx.safeSession.Session).(*vtgatepb.Session)
-	cloneCtx.safeSession = NewSafeSession(sess)
-	return cloneCtx
-}
